backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# AI Providers
anthropic==0.44.0
openai==1.59.0
google-generativeai==0.8.0
httpx==0.27.0

# Audio Processing
pydub==0.25.1

# Storage
boto3==1.34.0
python-magic==0.4.27

# WebSockets
websockets==12.0

# Utils
python-dotenv==1.0.0
uuid6==2024.1.12

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.27.0
